[7.x][ML] Make job data deletion reusable (#73602)
This commit moves the logic for deleting job data
from the `TransportDeleteJobAction` into the `JobDataDeleter`
class in order to allow reusing it. In particular, it will
be necessary for the implementation of the reset API.

Backport of #73602
